refactor: extract diff matcher
